dubai-final/api/requirements.txt
groq
flask
flask-cors
python-dotenv
gunicorn
resend
